requirements.txt
git+https://github.com/openai/whisper.git
ipywebrtc
openai
streamlit
plotly
wordcloud
streamlit-audiorecorder
st-audiorecorder